requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
python-multipart
bcrypt==5.0.0